Gradient Quad + Inspector
Modified colour script to change each pixel based on their place within the resolution ( gradient). and changed default filter (bilinear) to point based.

Created a script for a custom inspector. It draws the default inspector and checks for changes. If a change happened, it calls the FillTexture from our texture creating script.