Changed to URP.
Created a custom shader graph for the terrain that worked with the colour gradient on the mesh.

Created shader graph for the water. Started working with depth and trying to make the water look realistic. Ran into problems with using the camera as the view angle. So To Do: Make player first.